Change the default RefFFE
Remove now obsolete dateutil import.

Fix value being wrongfully set to None

Because all players have a defined `ffe_id`, the 'ffe_id' key points to
a None value. This means that the default value is never checked.
This commit simply allows the use of that default value.
Bonus points: it yields values in the same order of magnitude as Papi.