Merge back 2.8.0rc3 (#798)
* Release 2.8.0rc1 (#788)

* Release related fixes

* remove version change

* Update release notes

* Review + print documents

* Update player_admin_controller.py (#792)

* Restore error message on player search (#793)

* Restore error message on search

* Appy review

Co-authored-by: Pascal AUBRY <66429491+pascalaubry@users.noreply.github.com>

---------

Co-authored-by: Pascal AUBRY <66429491+pascalaubry@users.noreply.github.com>

* Release 2.8.0rc2

* Fix default FFE ID being None (#796)

* Change the default RefFFE

Remove now obsolete dateutil import.

Fix value being wrongfully set to None

Because all players have a defined `ffe_id`, the 'ffe_id' key points to
a None value. This means that the default value is never checked.
This commit simply allows the use of that default value.
Bonus points: it yields values in the same order of magnitude as Papi.

* Remove unused dateutil dependency

* Release 2.8.0rc3

---------

Co-authored-by: Molrn <76754000+Molrn@users.noreply.github.com>
Co-authored-by: Sammy Plat <sammy@sammyplat.fr>