Fix player movement
Before this commit, tournament players were deleted from the source
tournament before being added to the destination tournament.
However, when tournament players are deleted, the player is also
deleted.
This means that the player is not moved, unless they are in at least two
tournaments, which is a rare use case.

This commit fixes that by adding an event-level function to move players
between tournament